refactor/roomModel renamed userModel to roomModel
