+++ server/db/roomModel.js
@@ -1,0 +1,12 @@
+var mongoose = require('mongoose');
+
+var roomSchema = mongoose.Schema({
+    queue: [],
+    hash: String,
+    guests: [],
+    host: String,
+});
+
+var Room = mongoose.model('Room', roomSchema);
+
+module.exports = Room;
