--- server/db/userModel.js
@@ -1,10 +1,0 @@
-var mongoose = require('mongoose');
-
-var userSchema = mongoose.Schema({
-    queue: [],
-    hash: String
-});
-
-var User = mongoose.model('User', userSchema);
-
-module.exports = User;
